feat: add codegen for state and provider, move builder from main
