+++ lib/app/router/components/app_router_builder.dart
@@ -1,0 +1,27 @@
+// SPDX-License-Identifier: ice License 1.0
+
+import 'package:flutter/material.dart';
+import 'package:ion/app/components/global_notification_bar/global_notification_bar.dart';
+import 'package:ion/app/extensions/extensions.dart';
+
+class AppRouterBuilder extends StatelessWidget {
+  const AppRouterBuilder({super.key, this.child});
+
+  final Widget? child;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: SafeArea(
+        child: Column(
+          children: [
+            if (Theme.of(context).isInitialized) const GlobalNotificationBar(),
+            Expanded(
+              child: child ?? const SizedBox.shrink(),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
